changes to support staging decrypted file

+++ src/dataPipeline/steplibrary/PurgeLocationStep.py
@@ -2,7 +2,7 @@ import os
 import shutil
 from framework.pipeline import (PipelineStep, PipelineContext)
 from framework.enums import FilesystemType
-from framework.uri import FileSystemMapper
+from framework.uri import FileSystemMapper, native_path
 
 class PurgeLocationNativeStep(PipelineStep):
     def __init__(self, context_key: str='purge', **kwargs):
@@ -27,14 +27,14 @@ class PurgeLocationNativeStep(PipelineStep):
                 
             # This logic assumes a POSIX uri, make sure to map from whatever was specified to POSIX
             local_uri = FileSystemMapper.convert(uri, FilesystemType.dbfs)
-            if not local_uri.startswith('/dbfs'): local_uri = '/dbfs' + local_uri
+            local_uri = native_path(local_uri)
 
             self.logger.info(f'\tRemoving directory {local_uri}')
                 
             try:
-                #os.rmdir(mapped_uri)
-                shutil.rmtree(local_uri)
-                self.logger.info(f'\tSuccessfully removed {local_uri}')
+                if os.path.exists(local_uri):
+                    shutil.rmtree(local_uri)
+                    self.logger.info(f'\tSuccessfully removed {local_uri}')
 
             except Exception as e:
                 message = f'Failed to purge location {mapped_uri}'
